Extract win-rate result builders to keep file under 700-line limit
The thundering-herd defenses pushed domain-analysis-cache.ts past the
700-line hard limit. Extract `extractWinRates` and
`buildPairwiseWinRateModel` (pure data transformers used by
`buildDomainAnalysisResultFromSnapshot`) into a sibling
`domain-analysis-result-builders.ts`.
